changed handling of invalid bets
changed from text responses to react native alerts when invalid bet is
called

+++ src/screens/ScoresScreen.js
@@ -51,7 +51,7 @@ export default function ScoresScreen({navigation}) {
               />
               <Text>
                 {item['scores']['away']['total']}
-                {item['status']['short']}
+                {item['status']['long']}
                 {item['scores']['home']['total']}
               </Text>
               <Image
